+++ build.gradle
@@ -26,6 +26,9 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.boot:spring-boot-starter-oauth2-client'
+	implementation 'com.graphql-java-kickstart:graphql-spring-boot-starter:14.0.0'
+	implementation 'com.graphql-java-kickstart:graphql-java-tools'
+	implementation 'com.graphql-java-kickstart:graphql-java-servlet'
 	implementation 'org.springframework.boot:spring-boot-starter-websocket'
 	implementation 'org.springframework.security:spring-security-messaging'
 	compileOnly 'org.projectlombok:lombok'
