+++ src/main/java/Cook/Cookify_SpringBoot/domain/comment/repository/CommentRepository.java
@@ -1,6 +1,6 @@
 package Cook.Cookify_SpringBoot.domain.comment.repository;
 
-import Cook.Cookify_SpringBoot.domain.comment.Comment;
+import Cook.Cookify_SpringBoot.domain.comment.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
