+++ src/main/java/Cook/Cookify_SpringBoot/domain/comment/service/CommentService.java
@@ -1,7 +1,7 @@
 package Cook.Cookify_SpringBoot.domain.comment.service;
 
 
-import Cook.Cookify_SpringBoot.domain.comment.Comment;
+import Cook.Cookify_SpringBoot.domain.comment.entity.Comment;
 import Cook.Cookify_SpringBoot.domain.comment.dto.CommentRequestDto;
 import Cook.Cookify_SpringBoot.domain.comment.dto.CommentResponseDto;
 import Cook.Cookify_SpringBoot.domain.comment.exception.CommentException;
