+++ src/main/java/Cook/Cookify_SpringBoot/domain/heart/controller/HeartController.java
@@ -1,9 +1,7 @@
 package Cook.Cookify_SpringBoot.domain.heart.controller;
 
-import Cook.Cookify_SpringBoot.domain.heart.Heart;
 import Cook.Cookify_SpringBoot.domain.heart.dto.HeartRecipeDto;
 import Cook.Cookify_SpringBoot.domain.heart.service.HeartService;
-import Cook.Cookify_SpringBoot.domain.member.repository.GoogleMemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
