+++ src/main/java/Cook/Cookify_SpringBoot/domain/heart/repository/HeartRepository.java
@@ -1,8 +1,8 @@
 package Cook.Cookify_SpringBoot.domain.heart.repository;
 
-import Cook.Cookify_SpringBoot.domain.heart.Heart;
-import Cook.Cookify_SpringBoot.domain.member.GoogleMember;
-import Cook.Cookify_SpringBoot.domain.recipe.Recipe;
+import Cook.Cookify_SpringBoot.domain.heart.entity.Heart;
+import Cook.Cookify_SpringBoot.domain.member.entity.GoogleMember;
+import Cook.Cookify_SpringBoot.domain.recipe.entity.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
