+++ src/main/java/Cook/Cookify_SpringBoot/domain/heart/service/HeartService.java
@@ -1,14 +1,6 @@
 package Cook.Cookify_SpringBoot.domain.heart.service;
 
-import Cook.Cookify_SpringBoot.domain.heart.Heart;
 import Cook.Cookify_SpringBoot.domain.heart.dto.HeartRecipeDto;
-import Cook.Cookify_SpringBoot.domain.heart.repository.HeartRepository;
-import Cook.Cookify_SpringBoot.domain.member.GoogleMember;
-import Cook.Cookify_SpringBoot.domain.recipe.Recipe;
-import Cook.Cookify_SpringBoot.domain.recipe.repository.RecipeRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
