+++ src/main/java/Cook/Cookify_SpringBoot/domain/recipe/repository/RecipeDocsRepository.java
@@ -1,6 +1,6 @@
 package Cook.Cookify_SpringBoot.domain.recipe.repository;
 
-import Cook.Cookify_SpringBoot.domain.recipe.RecipeDocs;
+import Cook.Cookify_SpringBoot.domain.recipe.entity.RecipeDocs;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RecipeDocsRepository extends JpaRepository<RecipeDocs, Long> {
