+++ src/main/java/Cook/Cookify_SpringBoot/domain/recipe/repository/RecipeRepository.java
@@ -1,9 +1,7 @@
 package Cook.Cookify_SpringBoot.domain.recipe.repository;
 
-import Cook.Cookify_SpringBoot.domain.recipe.Recipe;
+import Cook.Cookify_SpringBoot.domain.recipe.entity.Recipe;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface RecipeRepository extends JpaRepository<Recipe, Long> {
 }
