+++ src/main/java/Cook/Cookify_SpringBoot/domain/recipe/service/RecipeService.java
@@ -3,14 +3,9 @@ package Cook.Cookify_SpringBoot.domain.recipe.service;
 import Cook.Cookify_SpringBoot.domain.recipe.dto.BriefRecipeDto;
 import Cook.Cookify_SpringBoot.domain.recipe.dto.RecipeDetailDto;
 import Cook.Cookify_SpringBoot.domain.recipe.dto.RecipeRequestDto;
-import Cook.Cookify_SpringBoot.domain.recipe.Recipe;
-import Cook.Cookify_SpringBoot.domain.recipe.repository.RecipeRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+import Cook.Cookify_SpringBoot.domain.recipe.entity.Recipe;
 
 import java.util.List;
-import java.util.Optional;
 
 
 public interface RecipeService {
